Move function for awaiting consensus sync with timeout upwards
